TODO: Must to fix http_error = 400
(HTTP) 400 Bad Request response status code indicates that the server cannot or will not process the request due to something that is perceived to be a client error (for example, malformed request syntax, invalid request message framing, or deceptive request routing).

Warning: The client should not repeat this request without modification.